Remove Emacs use-package and other Emacs config refactoring; Tooling updates

+++ tools/1_rs.sh
@@ -2,28 +2,38 @@
 
 log_header2 "Installing Rust tools..."
 
-function rustup {
-    if [ ! -d $CARGO_PATH ]; then
-        log_info "Rusting up..."
-        curl --proto '=https' --tlsv1.3 https://sh.rustup.rs -sSf | sh
+function install_rust {
+    log_info "Rusting up..."
+    if [ ! -d $CARGO_PATH ] || [[ -v DOTFILES_TOOLS_FORCE ]]; then
+        curl --proto '=https' --tlsv1.3 https://sh.rustup.rs -sSf | sh -s -- --no-modify-path
         log_info "Finished Rusting up."
     else
-        log_info "Already Rusted up."
+        log_info "Already Rusted up!"
     fi
 }
 
 function install_rust_tools {
-    rustup component add rustfmt
-    rustup component add rust-src
-    rustup component add clippy
-
-    # lsp
-    curl -L https://github.com/rust-lang/rust-analyzer/releases/latest/download/rust-analyzer-x86_64-unknown-linux-gnu.gz | gunzip -c - | sudo tee /usr/local/bin/rust-analyzer > /dev/null
-    sudo chmod +x /usr/local/bin/rust-analyzer
+    log_info "Installing Rust components..."
+    rustup component add rust-src > /dev/null 2>&1
+    rustup component add rustfmt > /dev/null 2>&1
+    rustup component add clippy > /dev/null 2>&1
+    log_info "Finished installing Rust components."
 }
 
-rustup
+function install_rust_lsp {
+    log_info "Installing Rust LSP..."
+    if ! [ -x $(command -v rust-analyzer) ] || [[ -v DOTFILES_TOOLS_FORCE ]]; then
+        curl -L https://github.com/rust-lang/rust-analyzer/releases/latest/download/rust-analyzer-x86_64-unknown-linux-gnu.gz | gunzip -c - | sudo tee /usr/local/bin/rust-analyzer > /dev/null
+        sudo chmod +x /usr/local/bin/rust-analyzer
+        log_info "Finished installing Rust LSP."
+    else
+        log_info "Rust LSP is already installed!"
+    fi
+}
+
+install_rust
 install_rust_tools
+install_rust_lsp
 
 log_header2 "Finished installing Rust tools.\n"
